eliminating duplicate models from apiserver, moving generated code to /internal
