+++ cmd/apiserver/handlers/handlers.go
@@ -8,14 +8,14 @@ import (
 	"github.com/microsoft/commercial-marketplace-offer-deploy/internal/data"
 )
 
-type HttpHandlerWithDatabase func(w http.ResponseWriter, r *http.Request, d *data.Database)
+type HttpHandlerWithDatabase func(w http.ResponseWriter, r *http.Request, d data.Database)
 
 // withDatabase wraps http handlers so a database is included as a func argument
 func WithDatabase(handler HttpHandlerWithDatabase) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		options := createOptionsFromConfiguration()
 		d := data.NewDatabase(options)
-		handler(w, r, &d)
+		handler(w, r, d)
 	}
 }
