--- cmd/apiserver/models/create_deployment.go
@@ -1,9 +1,0 @@
-package models
-
-type CreateDeployment struct {
-	Name string `json:"name"`
-
-	MultiStage bool `json:"multiStage,omitempty"`
-
-	Template *DeploymentTemplate `json:"template"`
-}
