--- cmd/apiserver/models/create_event_subscription.go
@@ -1,7 +1,0 @@
-package models
-
-type CreateEventSubscription struct {
-	Name string `json:"name,omitempty"`
-
-	Callback string `json:"callback,omitempty"`
-}
