--- cmd/apiserver/models/deployment_template.go
@@ -1,9 +1,0 @@
-package models
-
-type DeploymentTemplate struct {
-	Name string `json:"name,omitempty"`
-
-	Uri string `json:"uri,omitempty"`
-
-	Parameters []DeploymentTemplateParameter `json:"parameters,omitempty"`
-}
