--- cmd/apiserver/models/deployment_template_parameter.go
@@ -1,7 +1,0 @@
-package models
-
-type DeploymentTemplateParameter struct {
-	Name string `json:"name,omitempty"`
-
-	Value map[string]interface{} `json:"value,omitempty"`
-}
