--- cmd/apiserver/models/event.go
@@ -1,5 +1,0 @@
-package models
-
-type Event struct {
-	Name string `json:"name,omitempty"`
-}
