--- cmd/apiserver/models/invoke_deployment_operation.go
@@ -1,7 +1,0 @@
-package models
-
-type InvokeDeploymentOperation struct {
-	Name string `json:"name,omitempty"`
-
-	Parameters []OperationParameter `json:"parameters,omitempty"`
-}
