--- cmd/apiserver/models/invoked_operation_target.go
@@ -1,7 +1,0 @@
-package models
-
-type InvokedOperationTarget struct {
-	Type_ string `json:"type,omitempty"`
-
-	Id *InvokedOperationTargetId `json:"id,omitempty"`
-}
