--- cmd/apiserver/models/invoked_operation_target_id.go
@@ -1,7 +1,0 @@
-package models
-
-type InvokedOperationTargetId struct {
-	Type_ string `json:"type,omitempty"`
-
-	Value string `json:"value,omitempty"`
-}
