--- cmd/apiserver/models/operation_parameter.go
@@ -1,9 +1,0 @@
-package models
-
-type OperationParameter struct {
-	Name string `json:"name,omitempty"`
-
-	Value string `json:"value,omitempty"`
-
-	Type_ string `json:"type,omitempty"`
-}
