--- cmd/apiserver/models/operation_target_type.go
@@ -1,7 +1,0 @@
-package models
-
-type OperationTargetType struct {
-	ObjectType string `json:"objectType,omitempty"`
-
-	IdType string `json:"idType,omitempty"`
-}
